requirements.txt
PyQt5
yt-dlp
whisper 
openai-whisper 
pyannote.audio 
torch 
moviepy
certifi